Fixed missing `_instance` during deepcopy.
This fixes an issue where deep copies of `EmbeddedDocument` could
end up being invalid because the `_instance` field of the class
was not being serialized in the `__getstate__`/`__setstate__` calls
implemented by the parent class, `BaseDocument`.

The fix is to add `__getstate__`/`__setstate__` implementations to
`EmbeddedDocument` that directly serialize `_instance`, then defer
to the `super()` implementation for all other fields.

This also adds a unit test which verifies that `deepcopy` produces
usable `EmbeddedDocument` objects.